+++ examples/load_custom_dataset.py
@@ -23,7 +23,7 @@ file_path = os.path.expanduser('~/.surprise_data/ml-100k/ml-100k/u.data')
 reader = Reader(line_format='user item rating timestamp', sep='\t')
 
 data = Dataset.load_from_file(file_path, reader=reader)
-data.split(n_folds=5)
+data.split(n_folds=5)  # data can now be used normally
 
 # We'll use an algorithm that predicts baseline estimates.
 algo = BaselineOnly()